EVAKA-3899: refactor-decisions
