+++ frontend/packages/citizen-frontend/src/App.tsx
@@ -4,10 +4,10 @@
 
 import React from 'react'
 import { BrowserRouter, Route, Switch } from 'react-router-dom'
-import { Authentication } from './auth'
-import { Localization } from './localization'
-import Header from './header/Header'
-import Decisions from './decisions/Decisions'
+import { Authentication } from '~auth'
+import { Localization } from '~localization'
+import Header from '~header/Header'
+import Decisions from '~decisions/decisions-page/Decisions'
 import Footer from '~Footer'
 import DecisionResponseList from '~decisions/decision-response-page/DecisionResponseList'
 import GlobalErrorDialog from '~overlay/Error'
