+++ frontend/packages/citizen-frontend/src/decisions/PdfLink.tsx
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2017-2020 City of Espoo
+//
+// SPDX-License-Identifier: LGPL-2.1-or-later
+
 import InlineButton from '@evaka/lib-components/src/atoms/buttons/InlineButton'
 import { faFileAlt } from '@evaka/lib-icons'
 import { noop } from 'lodash'
