+++ frontend/packages/citizen-frontend/src/decisions/api.ts
@@ -1,9 +1,26 @@
+// SPDX-FileCopyrightText: 2017-2020 City of Espoo
+//
+// SPDX-License-Identifier: LGPL-2.1-or-later
+
 import { UUID } from '@evaka/lib-common/src/types'
 import { Failure, Result, Success } from '@evaka/lib-common/src/api'
 import { JsonOf } from '@evaka/lib-common/src/json'
-import { Decision } from '~decisions/types'
+import {
+  ApplicationDecisions,
+  Decision,
+  deserializeApplicationDecisions,
+  deserializeDecision
+} from '~decisions/types'
 import { client } from '~api-client'
 import LocalDate from '@evaka/lib-common/src/local-date'
+
+export async function getDecisions(): Promise<Result<ApplicationDecisions[]>> {
+  return client
+    .get<JsonOf<ApplicationDecisions[]>>('/citizen/decisions')
+    .then((res) => res.data.map(deserializeApplicationDecisions))
+    .then((data) => Success.of(data))
+    .catch((e) => Failure.fromError(e))
+}
 
 export async function getApplicationDecisions(
   applicationId: UUID
@@ -42,13 +59,4 @@ export async function rejectDecision(
     )
     .then(() => Success.of(undefined))
     .catch((e) => Failure.fromError(e))
-}
-
-function deserializeDecision(json: JsonOf<Decision>): Decision {
-  return {
-    ...json,
-    startDate: LocalDate.parseIso(json.startDate),
-    endDate: LocalDate.parseIso(json.endDate),
-    sentDate: LocalDate.parseIso(json.sentDate)
-  }
 }
