+++ frontend/packages/citizen-frontend/src/decisions/shared.tsx
@@ -1,6 +1,11 @@
+// SPDX-FileCopyrightText: 2017-2020 City of Espoo
+//
+// SPDX-License-Identifier: LGPL-2.1-or-later
+
 import { accentColors, greyscale } from '@evaka/lib-components/src/colors'
 import { faGavel, faCheck, faTimes } from '@evaka/lib-icons'
 import styled from 'styled-components'
+import { Decision, DecisionSummary } from '~decisions/types'
 
 export const Status = styled.span`
   text-transform: uppercase;
@@ -18,5 +23,20 @@ export const statusIcon = {
   REJECTED: {
     icon: faTimes,
     color: greyscale.lighter
+  }
+}
+
+type ComparableDecision = Decision | DecisionSummary
+
+export const decisionOrderComparator = (
+  decisionA: ComparableDecision,
+  decisionB: ComparableDecision
+) => {
+  if (decisionA.type === 'PRESCHOOL_DAYCARE') {
+    return 1
+  } else if (decisionB.type === 'PRESCHOOL_DAYCARE') {
+    return -1
+  } else {
+    return 0
   }
 }
